Adds hack instructions for those in a pinch (#73)
It is important for edge cases to stay edge cases. If someone is
invoking a builder, it is safe to assume they can manage the lifecycle
of what they are building. In the odd case they can't, it is better to
provide instructions vs absorb tech debt here.

This shows how to hack around lifecycle. It also fixes a subtle ordering
bug.