Also deploys Scala 2.13 on snapshot and release
The snapshot part will verify on merge to master. The syntax for deploy
a release should also work if master does. The only thing that will wait
for a release to find out is if the maven central sync works with two
versions of scala queued. It probably will work.